Changes around FabricPlayer
Replace FabricPlayer with FabricPlayerImpl, make FabricPlayer an interface and add DummyFabricPlayer as an enum singleton
